refactoring make credentials and add some sketch to get assertion
