+++ client/fido/fidocore.h
@@ -48,6 +48,9 @@ extern int FIDOCheckDERAndGetKey(uint8_t *der, size_t derLen, bool verbose, uint
 extern char *fido2GetCmdMemberDescription(uint8_t cmdCode, bool isResponse, int memberNum);
 extern char *fido2GetCmdErrorDescription(uint8_t errorCode);
 
+extern bool CheckrpIdHash(json_t *json, uint8_t *hash);
 extern int FIDO2CreateMakeCredentionalReq(json_t *root, uint8_t *data, size_t maxdatalen, size_t *datalen);
+extern int MakeCredentionalParseRes(json_t *root, uint8_t *data, size_t dataLen, bool verbose, bool verbose2, bool showCBOR, bool showDERTLV);
+extern int FIDO2CreateGetAssertionReq(json_t *root, uint8_t *data, size_t maxdatalen, size_t *datalen);
 
 #endif /* __FIDOCORE_H__ */
